requirements.txt
amqp==5.3.1
annotated-types==0.7.0
billiard==4.2.1
blinker==1.9.0
blis==0.7.11
catalogue==2.0.10
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.16.0
confection==0.1.5
cymem==2.0.11
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
filelock==3.18.0
Flask==3.1.1
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.33.2
idna==3.10
itsdangerous==2.2.0
jellyfish==1.2.0
Jinja2==3.1.6
joblib==1.5.1
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
lightgbm==4.6.0
lxml==6.0.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
nltk==3.9.1
numpy==1.26.4
packaging==25.0
pandas==2.3.0
pillow==11.3.0
preshed==3.0.10
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyMuPDF==1.26.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
segtok==1.5.11
sentence-transformers==5.0.0
six==1.17.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
sympy==1.14.0
tabulate==0.9.0
thinc==8.2.5
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.1
typer==0.9.4
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.1.3
xgboost==3.0.2
yake==0.6.0
